Restart Level Bug Fix -Fixed a bug where the game would crash if retrying/restarting the level multiple times. Presumably fixed by adding copy constructors? -Fixed the restart button leaving the menu up/the game paused.

+++ Game Engine/Character.cpp
@@ -2,7 +2,7 @@
 
 
 
-Character::Character(int h, int w, double positionX, double positionY, double velX, double velY, int hp) : Entity(h, w, positionX, positionY, velX, velY)
+Character::Character(int h, int w, double positionX, double positionY, double velX, double velY, double hp) : Entity(h, w, positionX, positionY, velX, velY)
 {
 	health = hp;
 }
