+++ Game Engine/Character.h
@@ -4,7 +4,7 @@ class Character :
 	public Entity
 {
 public:
-	Character(int h, int w, double positionX, double positionY, double velX, double velY, int hp);
+	Character(int h, int w, double positionX, double positionY, double velX, double velY, double hp);
 	virtual ~Character();
 	void hitWall(short direction) override;
 protected:
